Struct netcdf::file::MutableFile

pub struct MutableFile(_);

Mutable access to file

Implementations

impl MutableFile

pub fn root_mut(&mut self) -> Option<GroupMut>

Mutable access to the root group

Return None if this can't be a root group

pub fn variable_mut<'f>(&'f mut self, name: &str) -> Option<VariableMut<'f>>

Get a mutable variable from the group

pub fn variables_mut(&mut self) -> impl Iterator<Item = VariableMut>

Iterate over all variables in the root group, with mutable access

Examples

Use this to get multiple writable variables

let mut file = netcdf::append("file.nc")?;
let mut vars = file.variables_mut().collect::<Vec<_>>();
vars[0].put_value(1_u8, Some(&[2, 5]))?;
vars[1].put_value(1_u8, Some(&[5, 2]))?;

pub fn group_mut<'f>(&'f mut self, name: &str) -> Result<Option<GroupMut<'f>>>

Mutable access to subgroup

Errors

File does not support groups

pub fn groups_mut<'f>(
&'f mut self
) -> Result<impl Iterator<Item = GroupMut<'f>>>

Iterator over all groups (mutable access)

Errors

File does not support groups

pub fn add_attribute<'a, T>(
&'a mut self,
name: &str,
val: T
) -> Result<Attribute<'a>> where
T: Into<AttrValue>,

Add an attribute to the root group

pub fn add_dimension<'f>(
&'f mut self,
name: &str,
len: usize
) -> Result<Dimension<'f>>

Adds a dimension with the given name and size. A size of zero gives an unlimited dimension

pub fn add_unlimited_dimension(&mut self, name: &str) -> Result<Dimension>

Adds a dimension with unbounded size

pub fn add_group<'f>(&'f mut self, name: &str) -> Result<GroupMut<'f>>

Add an empty group to the dataset

pub fn add_variable<'f, T>(
&'f mut self,
name: &str,
dims: &[&str]
) -> Result<VariableMut<'f>> where
T: Numeric,

Create a Variable into the dataset, with no data written into it

Dimensions are identified using the name of the dimension, and will recurse upwards if not found in the current group.

pub fn add_variable_with_type<'f>(
&'f mut self,
name: &str,
dims: &[&str],
typ: &VariableType
) -> Result<VariableMut<'f>>

Create a variable with the specified type

pub fn add_opaque_type<'f>(
&'f mut self,
name: &str,
size: usize
) -> Result<OpaqueType>

Add an opaque datatype, with size bytes

pub fn add_vlen_type<'f, T: Numeric>(
&'f mut self,
name: &str
) -> Result<VlenType>

Add a variable length datatype

pub fn add_enum_type<'f, T: Numeric>(
&'f mut self,
name: &str,
mappings: &[(&str, T)]
) -> Result<EnumType>

Add an enum datatype

pub fn add_compound_type(&mut self, name: &str) -> Result<CompoundBuilder>

Build a compound type

pub fn add_string_variable<'f>(
&'f mut self,
name: &str,
dims: &[&str]
) -> Result<VariableMut<'f>>

Adds a variable with a basic type of string

pub fn add_variable_from_identifiers<'f, T>(
&'f mut self,
name: &str,
dims: &[Identifier]
) -> Result<VariableMut<'f>> where
T: Numeric,

Adds a variable from a set of unique identifiers, recursing upwards from the current group if necessary.