Struct netcdf::variable::VariableMut

pub struct VariableMut<'g>(_, _);

Mutable access to a variable.

Implementations

impl<'g> VariableMut<'g>

pub fn compression(&mut self, deflate_level: nc_type) -> Result<()>

Sets compression on the variable. Must be set before filling in data.

deflate_level can take a value 0..=9, with 0 being no compression (good for CPU bound tasks), and 9 providing the highest compression level (good for memory bound tasks)

Errors

Not a netcdf-4 file or deflate_level not valid

pub fn chunking(&mut self, chunksize: &[usize]) -> Result<()>

Set chunking for variable. Must be set before inserting data

Use this when reading or writing smaller units of the hypercube than the full dimensions lengths, to change how the variable is stored in the file. This has no effect on the memory order when reading/putting a buffer.

Errors

Not a netCDF-4 file or invalid chunksize

impl<'g> VariableMut<'g>

pub fn add_attribute<T>(&mut self, name: &str, val: T) -> Result<Attribute> where
T: Into<AttrValue>,

Adds an attribute to the variable

impl<'g> VariableMut<'g>

pub fn put_value<T: Numeric>(
&mut self,
value: T,
indices: Option<&[usize]>
) -> Result<()>

Put a single value at indices

pub fn put_string(
&mut self,
value: &str,
indices: Option<&[usize]>
) -> Result<()>

Internally converts to a CString, avoid using this function when performance is important

pub fn put_values<T: Numeric>(
&mut self,
values: &[T],
indices: Option<&[usize]>,
slice_len: Option<&[usize]>
) -> Result<()>

Put a slice of values at indices

pub fn put_values_strided<T: Numeric>(
&mut self,
values: &[T],
indices: Option<&[usize]>,
slice_len: Option<&[usize]>,
strides: &[isize]
) -> Result<usize>

Put a slice of values at indices, with destination strided

pub fn set_fill_value<T>(&mut self, fill_value: T) -> Result<()> where
T: Numeric,

Set a Fill Value

Errors

Not a netCDF-4 file, late define, fill_value has the wrong type

pub unsafe fn set_nofill(&mut self) -> Result<()>

Set the fill value to no value. Use this when wanting to avoid duplicate writes into empty variables.

Errors

Not a netCDF-4 file

Safety

Reading from this variable after having defined nofill will read potentially uninitialized data. Normally one will expect to find some filler value

pub fn endian(&mut self, e: Endianness) -> Result<()>

Set endianness of the variable. Must be set before inserting data

endian can take a Endianness value with Native being NC_ENDIAN_NATIVE (0), Little NC_ENDIAN_LITTLE (1), Big NC_ENDIAN_BIG (2)

Errors

Not a netCDF-4 file, late define

pub unsafe fn put_raw_values(
&mut self,
buf: &[u8],
start: &[usize],
count: &[usize]
) -> Result<()>

Get values of any type as bytes

Safety

When working with compound types, variable length arrays and strings create pointers from the buffer, and tries to copy memory from these locations. Compound types which does not have these elements will be safe to access, and can treat this function as safe

pub fn put_vlen<T: Numeric>(&mut self, vec: &[T], index: &[usize]) -> Result<()>

Get a vlen element