Trait netcdf::variable::Numeric

pub trait Numeric {
    fn from_variable(variable: &Variable) -> Result<Vec<Self>, String>
    where
        Self: Sized;

    fn read_variable_into_buffer(
        variable: &Variable, 
        buffer: &mut Vec<Self>
    ) -> Result<(), String>
    where
        Self: Sized;

    fn read_slice_into_buffer(
        variable: &Variable, 
        indices: &[usize], 
        slice_len: &[usize], 
        buffer: &mut Vec<Self>
    ) -> Result<(), String>
    where
        Self: Sized;

    fn slice_from_variable(
        variable: &Variable, 
        indices: &[usize], 
        slice_len: &[usize]
    ) -> Result<Vec<Self>, String>
    where
        Self: Sized;

    fn single_value_from_variable(
        variable: &Variable, 
        indices: &[usize]
    ) -> Result<Self, String>
    where
        Self: Sized;

    fn put_value_at(
        variable: &mut Variable, 
        indices: &[usize], 
        value: Self
    ) -> Result<(), String>
    where
        Self: Sized;

    fn put_values_at(
        variable: &mut Variable, 
        indices: &[usize], 
        slice_len: &[usize], 
        values: &[Self]
    ) -> Result<(), String>
    where
        Self: Sized;

    fn as_void_ptr(&self) -> *const c_void;
}

This trait allow an implicit cast when fetching a netCDF variable

Required Methods

fn from_variable(variable: &Variable) -> Result<Vec<Self>, String> where
Self: Sized,

Returns the whole variable as Vec

fn read_variable_into_buffer(
variable: &Variable,
buffer: &mut Vec<Self>
) -> Result<(), String> where
Self: Sized,

Read the variable into a buffer and update its length.

fn read_slice_into_buffer(
variable: &Variable,
indices: &[usize],
slice_len: &[usize],
buffer: &mut Vec<Self>
) -> Result<(), String> where
Self: Sized,

Read a slice of a variable into a buffer and update its length.

fn slice_from_variable(
variable: &Variable,
indices: &[usize],
slice_len: &[usize]
) -> Result<Vec<Self>, String> where
Self: Sized,

Returns a slice of the variable as Vec

fn single_value_from_variable(
variable: &Variable,
indices: &[usize]
) -> Result<Self, String> where
Self: Sized,

Returns a single indexed value of the variable as Self

fn put_value_at(
variable: &mut Variable,
indices: &[usize],
value: Self
) -> Result<(), String> where
Self: Sized,

Put a single value into a netCDF variable

fn put_values_at(
variable: &mut Variable,
indices: &[usize],
slice_len: &[usize],
values: &[Self]
) -> Result<(), String> where
Self: Sized,

put a SLICE of values into a netCDF variable at the given index

fn as_void_ptr(&self) -> *const c_void

Returns self as a C (void *) pointer